requirements.txt
feedparser==6.0.11
requests==2.31.0
transformers==4.52.4
torch==2.6.0
streamlit==1.32.0
python-dotenv==1.0.0
pydub==0.25.1 